Fix config checking, promote hw-health to mandatory + more
Fix config checking to use 'charm' instead of 'charm-name' which
doesn't exist in bundles.

Demote passing config checks to debug (rather than info) log level.

If a config variable is 'eq' but we don't see it, warn the user.

Promote hw-health to a mandatory subordinate for hosts (will false
positive on VMs for now).

Promote logrotate to a mandatory subordinate everywhere.

Add trilio charms as known.