config: deprecate option --dump-state
Using two different command line arguments for the same feature was a
bit confusing. Especially the command line option --dump-state was used
as a boolean flag however it expected a no-op argument to be passed. The
new behavior will gladly accept any number of arguments including zero
and do nothing with them except for printing a deprecation notice.

Dumping the cloud state will be activated by the --output-dir option
only.

We move from this usage:
  $ juju-lint --dump-state NOOP_ARG --output-dir DIR ...

to this:
  $ juju-lint --output-dir DIR ...

Also during the investigation of this issue, another issue was observed.
This option suggests in its help text that the lint reports are also
going to be written to the output directory, however in the resulting
output they were absent. Another launchpad issue is filed about this and
the "lint reports" part is taken out of the help text for now, to not to
confuse the user.

This change also adds a check for the existence and write access to the
directory before running any linting.

Closes-bug: #1987951
Signed-off-by: Mert Kırpıcı <mert.kirpici@canonical.com>