tox: separate requirements
The idea behind this change is to separate the requirements to avoid
installing redundant packages for each different testing/linting
environment that we are using during a regular development cycle.

Also now we are getting rid of preparing the dev-environment and
installing pre-commit via virtualenv and moving this operation inside
tox.ini which all the other virtual environments are residing, relieving
the Makefile of this duty.

After the renaming process in 30066d3b, makefile dev-environment target
was overlooked, ending up with the broken target. This change is fixing
that also.

Since this issue[0] is resolved, unpinned the flake8 version.

The .venv/ created by dev-environment was not ignored by git, fixed
that.

[0] https://github.com/csachs/pyproject-flake8/issues/13

Closes-bug: 1989575
Signed-off-by: Mert Kırpıcı <mert.kirpici@canonical.com>